+++ RMIS.Domain/RMIS.Domain/RiceMill/BagStockInfoEntity.cs
@@ -17,8 +17,8 @@ namespace RMIS.Domain.RiceMill
         public string MRiceBrandID { get; set; }
         public string VehicalNo { get; set; }
         public string DriverName { get; set; }
-        public Int16 TotalBags { get; set; }
-        public decimal Price { get; set; }
+        public int TotalBags { get; set; }
+        public double Price { get; set; }
         public DateTime PurchaseDate { get; set; }
         #region Methods
         /// <summary>Determines whether the specified object is equal to this instance.</summary>
